# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mirth Connect API",
    "version": "3.9.0",
    "description": "Node.js Mirth Connect REST API — compatible with Mirth Connect Administrator.\n\nThis API provides full channel management, message processing, user authentication, and monitoring capabilities. All endpoints (except health checks and login) require authentication.",
    "license": {
      "name": "MPL-2.0",
      "url": "https://www.mozilla.org/en-US/MPL/2.0/"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Local development"
    },
    {
      "url": "http://localhost:8080",
      "description": "Default production port"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID",
        "description": "Session cookie set after login via POST /api/users/_login"
      },
      "sessionHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-ID",
        "description": "Alternative session ID passed as header"
      }
    },
    "schemas": {
...
      "LoginStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SUCCESS",
              "SUCCESS_GRACE_PERIOD",
              "FAIL",
              "FAIL_LOCKED_OUT",
              "FAIL_EXPIRED"
            ]
          },
          "message": {
            "type": "string"
          },
          "updatedUsername": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or short message"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message"
          }
        },
        "required": [
          "error"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "User ID (auto-generated)"
          },
          "username": {
            "type": "string",
            "description": "Unique username"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "manager",
              "operator",
              "monitor"
            ],
            "description": "User role for RBAC"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phoneNumber": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp"
          },
          "gracePeriodStart": {
            "type": "string",
            "format": "date-time"
          },
          "strikeCount": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "CreateUser": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "description": "Defaults to \"admin\" if omitted"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "manager",
              "operator",
              "monitor"
            ]
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phoneNumber": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          }
        },
        "required": [
          "username"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/users/_login": {
      "post": {
        "summary": "Login",
        "description": "Authenticate with username and password. Returns session cookie.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login result (check status field for SUCCESS/FAIL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginStatus"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or locked account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginStatus"
                }
              }
            }
          },
          "429": {
            "description": "Too many login attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/_logout": {
      "post": {
        "summary": "Logout",
        "description": "Destroy current session and clear session cookie.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logged out successfully"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "User list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Username is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/current": {
      "get": {
        "summary": "Get current user",
        "description": "Returns the currently authenticated user.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userIdOrName}": {
      "get": {
        "summary": "Get user by ID or username",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "User ID (integer) or username (string)"
            },
            "required": true,
            "description": "User ID (integer) or username (string)",
            "name": "userIdOrName",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}": {
      "put": {
        "summary": "Update user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "User ID"
            },
            "required": true,
            "description": "User ID",
            "name": "userId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "User updated"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "User ID"
            },
            "required": true,
            "description": "User ID",
            "name": "userId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/password": {
      "put": {
        "summary": "Update user password",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "User ID"
            },
            "required": true,
            "description": "User ID",
            "name": "userId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Empty array if valid, or array of requirement violations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Password requirements not met",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/loggedIn": {
      "get": {
        "summary": "Check if user is logged in",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "User ID"
            },
            "required": true,
            "description": "User ID",
            "name": "userId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Login status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/_checkPassword": {
      "post": {
        "summary": "Check password against requirements",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Array of requirement violations (empty if valid)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/preferences": {
      "get": {
        "summary": "Get user preferences",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Preference map",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update user preferences",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "sessionHeader": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "userId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Preferences updated"
          }
        }
      }
    },
...
  },
...
}